pyproject.toml
[tool.poetry]
name = "fmb-experiment"
version = "0.1.0"
description = ""
authors = ["Ashley Felton <ashley@ropable.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.31"
bottle = "^0.12.19"
cx-Oracle = "^8.3.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
ipython = "^8.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
